requirements.txt
aws-cdk-lib>=2.158.0,<3.0.0
constructs>=10.3.0,<11.0.0
boto3>=1.35.0
pydantic>=2.8.0
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
aws-lambda-powertools>=3.0.0
python-json-logger>=2.0.7
typing-extensions>=4.12.0
moto>=5.0.0
pytest>=8.2.0
pytest-asyncio>=0.23.6
